requirements.txt
streamlit==1.56.0
ultralytics==8.4.37
opencv-python==4.13.0.92
plotly==6.7.0
google-generativeai==0.8.6
python-dotenv==1.2.2
